create single_client_workload_config
