+++ crates/walrus-stress/src/single_client_workload.rs
@@ -3,4 +3,17 @@
 
 //! Single client workload.
 
-pub(crate) mod single_client_workload_args;
+use single_client_workload_config::{
+    RequestTypeDistributionConfig,
+    SizeDistributionConfig,
+    StoreLengthDistributionConfig,
+};
+
+pub(crate) mod single_client_workload_arg;
+pub(crate) mod single_client_workload_config;
+
+pub struct SingleClientWorkload {
+    data_size_config: SizeDistributionConfig,
+    store_length_config: StoreLengthDistributionConfig,
+    request_type_distribution: RequestTypeDistributionConfig,
+}
